\name{dep_installed}
\alias{dep_installed}
\title{Check that the installed version of a package satisfies the requirements}
\usage{
  dep_installed(dep_name, dep_compare = NA, dep_ver = NA)
}
\arguments{
  \item{dep_name}{The name of the package with objects to
  import}

  \item{dep_ver}{The version of the package}

  \item{dep_compare}{The comparison operator to use to
  check the version}
}
\description{
  Check that the installed version of a package satisfies
  the requirements
}
\keyword{internal}

