\name{sources}
\alias{sources}
\title{A list of sources.}
\usage{
  sources(..., .sources = list())
}
\arguments{
  \item{...}{individual source objects}

  \item{.sources}{a list of source objects}
}
\description{
  A list of sources.
}
\seealso{
  \code{\link{default_sources()}} for a list of sources
  generated to represent a useful set of resources.
}
\keyword{internal}

